Move value extraction logic to Applier
